exercise 04 done
made a function to remove multiple values from an array.
used an array function .filter() to check if each item
should be passed to the new array. so in the value
passed to .filter() is a callback function to iterate
through the given array and return true if it doesnt
match with all the target values.